+++ src/main/java/baseball/validation/AskValidation.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.validation;
 
 import baseball.resource.AskRestartValue;
