+++ src/main/java/baseball/validation/GameValidation.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.validation;
 
 import baseball.resource.GameValue;
